Change TLS-SNI to require hostname not specified by validator
This replaces tls-sni-01 with tls-sni-02, which requires that domains
being validated return a certificate which contains a hostname that is
not revealed by the server in its handshake. This ensures that
challenges cannot be passed by domains which auto-generate self-signed
certificates in response to requests with SANs for domains not
previously seen.