Add metadata to directory
This commit does two things: Firstly, it adds a metadata section to the
directory, allowing terms of service and website URL to be stored, as
well as a list of recognised CAA issuer identifiers.

Secondly, it adds a Link (rel="directory") to all resources pointing
back to the directory. This ensures that given the URL of an ACME
resource, the URL of the directory needed to perform operations relating
to it can be found.

Fixes #59.