Remove combinations array.
This is a fairly complicated part of the protocol, and not used in practice. For
instance, in Let's Encrypt's implementation, there are always three challenges,
any one of which may be fulfilled by the client.

After this change, all challenges are considered to be combined with an "OR."
That is, any challenge within an authorization may be completed to make the
authorization valid.

Authorizations within the new-application object are considered to be combined
with an "AND." That is, all of them must become valid before the certificate
will be issued. The combination of the two means that we have similar
expressiveness as before, even without the combinations array.