Clarify "new-X" resources paragraph. (#179)
Prior to this commit the specification seems to be contradictory,
indicating that the "new-X" resources must not be duplicated, but that
they can live at different URIs. It's not clear to me what this MUST is
trying to prevent, especially in light of an allowance of multiple URIs
if functionality is equivalent.

Similarly, while the text prior to this commit says "new-X" I believe
the intentions of the statements that follow would include some of the
resources that have been added since, e.g. "key-change".

This commit updates the "new-X" language to include the additional
singular resources and removes the MUST.