Specify account by kid (reg URL) rather than key.
Previously, each authenticated POST included a copy of its public key in the JWS
protected header. Servers were required to canonicalize this key in some way and
look it up in a database, then verify the payload. However, servers also had the
option of verifying the payload, and *then* looking up the key in the database.
This opened up a category of potential bugs based on truncated database fields
or hash collisions, where an attacker could craft a key pair the would
potentially be matched to someone else's account.

Incidental benefits: Switching to kid makes payloads somewhat smaller, and
removes the need to define key equality in the ACME spec.