Standardizes on "proactive" certificate issuance.
Based on the acme mailing list discussion[0] on "proactive" vs
"on-finalization" issuance it seems there is more support for the
proactive approach.

Speaking on behalf of Let's Encrypt and the Boulder developers we're
willing to compromise and support proactive issuance as the sole
issuance method in order to simplify the protocol & implementations.

This commit changes the language around certificate issuance to reflect
that the server MUST issue the certificate proactively once all the
required challenges for an application have been fulfilled. The open
issue corresponding to making this decision is removed now that there is
one supported option.

In an effort to close out open questions around applications this commit
also removes the open issue around multiple certificates per
application. If there is a demand for automating repeat issuance there
should be a new thread started with a concrete proposal.

0: https://mailarchive.ietf.org/arch/msg/acme/0lVmGl8e-rmSH0x7ycDW5dj6GAY