Simplifies Section 6.1.2 - "Registration Objects".
In Section 6.1.2 "Registration Objects" the "applications" field of
the registration resource is described in great detail. This is
confusing for two reasons:
  1) both the text and the example registration object are still using
     the "authorization" term that applications replace.
  2) the text is written as if the applications should be directly
     embedded in the registration object returned.

This commit addresses both issues by describing the "applications" field
as an array of application URIs that the client can fetch. The example
"registration" object is updated accordingly.

The note indicating a server SHOULD include pending applications and
SHOULD NOT include invalid applications remains as-is. The note
indicating a server MAY return an incomplete list is removed - it is
more complicated and with only URIs being returned, likely an unneeded
optimization.