Default to PEM with chain for certificates.
One of the most common ACME deployment failures observed in practice is for
servers to be configured to serve only the end-entity certificate, without the
intermediate certificates. This is a particularly pernicious problem because
some browsers will still trust the resulting one-certificate chain, due to
caching or fetching of URLs from Authority Information Access. But other
browsers will not, resulting in a "works on my computer" problem.

Arguably this configuration is the result of incorrect clients, but we should
expect that most clients will do the easiest thing. This change aligns the
easiest thing with the most correct thing.