Restores "next" semantics to applications list.
This commit addressed feedback provided by @jsha on the ACME list. In
particular it restores the idea that the registration object should
return a single URI from which applications can be retrieved, and
further, that the "applications list" URI may return partial results
relying on pagination from a "Link" header.