Updates Section 6.1.2 - "Registration Objects" for Applications. (#192)
* Simplifies Section 6.1.2 - "Registration Objects".

In Section 6.1.2 "Registration Objects" the "applications" field of
the registration resource is described in great detail. This is
confusing for two reasons:
  1) both the text and the example registration object are still using
     the "authorization" term that applications replace.
  2) the text is written as if the applications should be directly
     embedded in the registration object returned.

This commit addresses both issues by describing the "applications" field
as an array of application URIs that the client can fetch. The example
"registration" object is updated accordingly.

The note indicating a server SHOULD include pending applications and
SHOULD NOT include invalid applications remains as-is. The note
indicating a server MAY return an incomplete list is removed - it is
more complicated and with only URIs being returned, likely an unneeded
optimization.

* Updates section 6.3 "Registration" for correct fields.

Section 6.3 "Registration" describes three fields that must be ignored
by the server in a registration body: key, authorizations and
certificates. This should be "key" and "applications". Certificates is
not specified and authorizations have been updated to applications.

* Restores "next" semantics to applications list.

This commit addressed feedback provided by @jsha on the ACME list. In
particular it restores the idea that the registration object should
return a single URI from which applications can be retrieved, and
further, that the "applications list" URI may return partial results
relying on pagination from a "Link" header.

* Moves `Applications List` subsection.

Per review-feedback, this commit moves the `Applications List`
subsection to the `Registration Objects` section instead of `Application
Objects`.

* Verisimilitude++

* Fixes header levels