Clarify initial section.
Several places talked about account key pairs as being the primary notion of an
account, while the account is now primary and has a public key associated with
it. Updated to clarify that.

Clarified that the account request can contain more than just contact
information.

Fixed the order of operations in protocol summary to reflect new-order flow.