More editorial fixes.
Include "processing" as a possible status for an order (this was listed
elsewhere but not consistent).

Include a tls-sni-02 challenge in example authz.

Clarify that finalizing an authorization involves *modifying* the expires field,
not adding it.

Clarify in revocation section the two types of accounts that can request a
certificate revocation.

Expand examples in challenge section to match format in the rest of the doc,
including HTTP headers and JWS wrapping.

Indicate for HTTP, TLS SNI, and DNS, that making multiple requests is not only
supported, but RECOMMENDED.

Fix a mismatch between "client configurable" in description of registries and
"configurable" in the table of things to be registered.