review: reorder sentence to address comma confusion and disagreement between conjunction and negation
There are a number of changes here.
The original structure was:
B should do K, and should not do L in Z

Normally you would write:
C v P and Q in Y
or possibly
C v P and v Q in Y
or possibly
C v P, but not v Q, in Y

* and should not be used with not
* comma for exclusion should be wrapped around the aside

Alternatively, for clarity (this is an aside), one can do:
In Z, B should do K and L
or
In Z, B should do K but not L
this last variation is the form I propose here.

Note: I've changed SHOULD NOT include to SHOULD exclude,
you can swap that back, it doesn't change the spirit of 'but'.