Disambiguate "server" where required.
As pointed out by @jsoref on the ACME mailing list there are a few
places where there was some potential ambiguity about what "server"
meant.

This commit is the result of a once-through pass looking for places
where we could be more specific when saying "server". In most cases the
default definition of "server" from the Terminology section applied and
no change was made. I attempted to clarify any usage of "server" that
wasn't qualified sufficiently to differentiate from the ACME server.