Remove extraneous link relations and change 'directory' to 'index' (#250)
As they are all fulfilled by the `directory` endpoint and are unnecessary. The `ct-sct` relation is removed as the SCT will be provided by the CA via the certificate or a OCSP response, if it doesn't provide these then the client should be able to retrieve the SCT itself from the logs it expects to find the certificate in (or submit it itself).

Also removes a paragraph about Certificate Transparency from the 'Downloading the Certificate' section as it is irrelevant to the section and discusses operations of the CA unrelated to the ACME protocol.

Additionally changes the relation `rate-limit` which is not registered with a URI extension using the ACME URN namespace `urn:ietf:params:acme:documentation`.