Remove new-authz fields "existing" and "accept". (#275)
I believe they are unnecessary. The server can set its own policy on whether to
give back an existing or a new authz, and the client can choose to accept or
reject that.

Hugo also pointed out on-list that the status codes for these are inconsistent
with the status codes for new-account. If we keep these, we'd want to fix that.
But simpler to delete, I think.