Normalizes 'revoke-cert', 'account'.
Fixes per feedback from Niklas Keller on the mailing list.

Section 6.1 referenced "revoke-certificate" where elsewhere using
"revoke-cert", this commit replaces it with "revoke-cert".

Section 7.1 has a resource diagram that previously labelled the
termination of the "new-account" flow as "acct", this commit replaces
that label with "account". Elsewhere the use of "acct" is limited to
URLs and not referencing specific resources/spec content.