Rewords token entropy justification.
Per feedback from Zach Shepherd[0] on the mailing list the justification
for the 128 bits of required token entropy from section 8.2 is at odds
with the actual rationale (as I understand it).

This commit updates the justification for the entropy requirement to not
mention being secret from an attacker, but instead not predictable by
a naive challenge response server that would complete challenges
without having been a party to the authz/chal creation.