Adds "unsupportedContact" error, makes "mailto:" MTI.
Firstly, this commit updates the table of errors in Section 6.6
"Errors" to add an "unsupportedContact" error type in addition to
"invalidContact". The former is returned when the contact URI protocol
is unsupported and the latter when the protocol is supported but the
provided contact is invalid.

Secondly, this commit updates Section 7.3 "Account Creation" to both
mention the new "unsupportedContact" error type and to make support of
the "mailto:" protocol a MUST.