Additional editorial changes (#285)
* Editorial: use 'header parameter' to refer to JWS header parameters

Some sections of the document were consistent in their use of 'header
parameter'. Others used 'header', 'parameter', 'header parameter', and
'field' seemingly-interchangably. This standardizes on 'header parameter'
to avoid ambiguity.

* Editorial: Fix typo in section reference

* Editorial: avoid ambiguous use of 'administratively'

"Client" and "sever" are defined in the terminology section. "User" and
"administratively" have not been defined and their use may lead to confusion.

* Editorial: use parallel sentence structure in section 7.1.4