Adds "unsupportedContact" error, makes "mailto:" MTI. (#291)
* Adds "unsupportedContact" error, makes "mailto:" MTI.

Firstly, this commit updates the table of errors in Section 6.6
"Errors" to add an "unsupportedContact" error type in addition to
"invalidContact". The former is returned when the contact URI protocol
is unsupported and the latter when the protocol is supported but the
provided contact is invalid.

Secondly, this commit updates Section 7.3 "Account Creation" to both
mention the new "unsupportedContact" error type and to make support of
the "mailto:" protocol a MUST.

* Clarifies that "A" contact is in error, not "The".

* Addresses editorial feedback.

Updates "URL" to "URI". Refers to the "mailto scheme" instead of
protocol. Clarifies the requirement that the server MUST support the
mailto scheme for account URIs.

* Clarifies contact validation paragraph

* Quotes around "mailto"