Protect HTTP validation from stateless clients
As discussed on the mailing list[1], a stateless client can be written
for http-01 because the token is provided by the server as a part of
the validation request.

This change introduces http-07, which uses a SHA-256 digest of the token
in the request instead of the token itself, as suggested by Logan Widick.

Because the expected request body is unchanged between http-01 and
http-07, clients and servers could trivially support both during a
transition period.

Note: I think it may make more sense to call this http-02, especially if
      the version numbers in challanges are not going to be removed in
      the final version of the RFC, but this change follows the  current
      versioning strategy as described in the Editor's Note at the end
      of Section 8.

1 - https://www.ietf.org/mail-archive/web/acme/current/threads.html#01941