Clarify that the "newAuthz" directory endpoint is optional.
In Section 7.4.1 "Pre-Authorization" the spec says:

     If a CA wishes to allow pre-authorization within ACME, it can offer
     a "new authorization" resource in its directory by adding the field
     "newAuthz" with a URL for the new authorization resource.

That text indicates that the CA may wish to *not* support
pre-authorization in which case the "newAuthz" resource will not be
present.

This commit raises that possibility in Section 7.1.1 "Directory" where
a developer is likely to look to understand which resources must be
supported/present in the directory and which are optional ("meta",
"newAuthz").