Add "wildcard" field to authorizations. (#411)
`NewOrder` requests are allowed to send DNS type identifiers with
a value that contains a wildcard prefix ('*') as the leftmost label.

The authorization object created for the wildcard identifiers in the
requests are **not** allowed to carry the wildcard prefix in the
authorization identifier per spec language:
> An authorization returned by the server for a wildcard domain name
> identifier MUST NOT include the asterisk and full stop ("\*.")
> prefix in the authorization identifier value

This is sensible because the authorization is not for the logical
wildcard but for the specific base domain.

This leaves an ambiguous situation for client developers/users.
Imagine a situation where a user requests a new order with the following
request body:

```
{
    "identifiers": [
        { "type": "dns", "value": "example.com" },
        { "type": "dns", "value": "*.example.com" }
    ]
}
```

It's possible two authorizations will be created for this case, but both
will contain the identifier `{ "type":"dns", "value": "example.com"}`.
This can be confusing to users! (e.g.  shred/acme4j#53)

If one of the authorizations fails a challenge validation operation the
client can't know whether it was the "example.com" validation, or the
"*.example.com" validation that failed and won't be able to prompt the
user to retry with the failed domain removed. Generally speaking it
makes presenting accurate error information difficult without
over-fitting Let's Encrypt's particular Wildcard issuance policies.
(E.g. see discussion in
https://github.com/certbot/certbot/pull/5620#issuecomment-369111947 and
https://github.com/certbot/certbot/issues/5613#issuecomment-368097551)

This commit introduces an optional boolean authorization field that can
be used to indicate the authorization's identifier should be considered
the base domain of a wildcard identifier that was requested.