Fix formatting in IANA section.
For pem-certificate-chain, the "Addition information" section was
getting jumbled all onto one line by the Markdown to HTML generator.

The table for Fields in Account Objects was jumbled.

In the template for some registries, a field was called "Client
configurable" in the next but "Configurable" in the table. Changed to
always say "Configurable."